+++ R/1_model_parameters.R
@@ -441,6 +441,11 @@ model_parameters.default <- function(model,
                                      ...) {
   dots <- list(...)
 
+  # sanity check, warn if unsupported argument is used.
+  if (verbose) {
+    .check_dots(dots = dots, not_allowed = c("include_sigma", "wb_component"), class(model)[1])
+  }
+
   out <- tryCatch(
     {
       .model_parameters_generic(