+++ R/methods_survey.R
@@ -12,7 +12,7 @@ model_parameters.svyglm <- function(model,
                                     verbose = TRUE,
                                     ...) {
   if (insight::n_obs(model) > 1e4 && ci_method == "likelihood") {
-    message(insight::format_message("Likelihood confidence intervals may take longer time to compute. Use 'df_method=\"wald\"' for faster computation of CIs."))
+    message(insight::format_message("Likelihood confidence intervals may take longer time to compute. Use 'ci_method=\"wald\"' for faster computation of CIs."))
   }
 
   out <- .model_parameters_generic(